Add collision detection demo based on generated terrain
The current algorithm is:
1. A posteriori collision detection
2. Sphere - Plane collision detection

The Object class simulates sphere based on the
CD(collision detection) impl, upon each frame, the object finds
the one triangle plane beneath it in the terrain by traversing Node
structure, and detects if the Object sphere collide with the
triangle plane. If collision is detected, bounce is performed.

Signed-off-by: plysan <plysab@gmail.com>