Fix jittering problem
The terraini's vertex was generated with earth's core in the center
of the model coordinate, which cause the terrain far from origin
point.
But opengl only use floating point which causes precision lose if
the camera is around the terrain polygons.

The solusion is generate the terrain's vertex and move then until
the viewing position of the terrain is at the center of the model
coordinate. This proccess performs in double precision using CPU.
Then convert them back to floating point precision. In this
way, vertex near the viewing position has enough precision so that
not to jitter.

Signed-off-by: plysan <plysab@gmail.com>