Support non-square quardtree node in integer degree scale
Means the quardtree node is not required to be square if the size
of node is integer degrees. node that is less than one degree is
still required to be square node, they are not necessary to be
non-square node now, because dem sources i'm using is based on
integer degree. This is a improvement for patch:
a81f054b8f5703511d61cbddd2db3080f26c390c

Signed-off-by: plysan <plysab@gmail.com>