Only update binding of GL_ELEMENT_ARRAY_BUFFER on use-buffer step
Before this patch, the two buffer object for element array is
always binded to GL_ELEMENT_ARRAY_BUFFER for no reason (I was
newbie in opengl :X ). This means, the updating buffer object will
bind to GL_ELEMENT_ARRAY_BUFFER when it is mapped and unmapped

In this patch, one of the two buffer object for element array is
binded to GL_ARRAY_BUFFER when it is in map-buffer, update-buffer,
unmap-buffer steps, and the other one is binded to
GL_ELEMENT_ARRAY_BUFFER when the buffer object is in use-buffer
step.

In this way, opengl doesn't have to switch the using element array
(the buffer object which is binded to GL_ELEMENT_ARRAY_BUFFER),
while the other element buffer object is binded to GL_ARRAY_BUFFER
for update(map and unmap), and this is good for performance

steps is referenced from: 3260cf42b0804d5648a8f6e786890071e54bc101

Signed-off-by: plysan <plysab@gmail.com>