Good practise for memory management of buffer object data
The heap memory allocation and deletion of buffer objects should
not be done by the algorithm implementation, but by the main
drawing program(or maybe by some seperate memory manager code
blocks in the future), so that the main program can have good
knowledge and control of the heap memory.

In this way, algorithm implementations just need to tell main
program how much memory is needed, and the main program takes care
of the memory management. This practise should be obeyed in future
buffer object data generation implementations.

Signed-off-by: plysan <plysab@gmail.com>