Add skydome's vertex and element data
The skydome's geometry is placed on top boundary of the atmosphere.

skydome and terrain are going to share the same buffer. Currently
only vertex and element data are generated, the coloring of the sky
will use 3D scattering texture in future commits.

During the proccess, I decided to change constants.hpp to vars.hpp
for the plan to store all the variables used in multiple files.

Signed-off-by: plysan <plysab@gmail.com>