Make skydome better shape
Skydome is generated using sphere coordinates, but with the center
point calculated from viewer's coordinate. This will have
singularity problem when skydome is over poles.

The solution here is generate skydome with the center point
calculatd from north pole's coordinate, and then rotate the center
of the whole skydome to viewer's coordinate.

Signed-off-by: plysan <plysab@gmail.com>