Support atmospheric scattering effect for outer space and terrain
When in space, camera is virtually moved to the top boundary of
atmosphere along the viewing direction, and scattering texture is
fetched from there.

To render the terrain, camera is virtually moved to the terrain's
face along the viewing direction, and scattering color is fetched
from there and subtracted by the scattering color on the top
boundary of atmosphere on the viewing direction.

A uniform "render_target" is introduced to distinguish rendering
shader algorithm of terrain and sky.

Signed-off-by: plysan <plysab@gmail.com>