Add space cube to the background
I don't have plan to go out of solar system in the near future.

Because the cube's vertex amount is little, I just put it in the
existing buffer. And for that, I planned a static area in the array
buffers to store static drawing data such as space cube. If the
static data gets big as time goes by, I may consider making a
seperate array buffer.

Signed-off-by: plysan <plysab@gmail.com>