Correct calculation of light scattering from camera to terrain
Scaterring color for the terrain is not stored in the scatter
texture, since scatter texture only contains scatter color info
staring from sealevel.

To get scatter color for arbitrary terrain, I replaced the alpha
channel of scatter texture with optical depth, representing the
optical depth of each point&direction the scatter texture
describes. And then the color is calculated in the following way:

Assume:
* I_a and I_b is scatter colors fetched from camera position
  and terrain vertex position,
* D_a and D_b is optical depth fetched from the positoins of I_a
  and I_b,
* I_o is the result color(scatter color from camera to terrain),
we have:

I_o = I_a - I_b * optical_depth_from_a_to_b
    = I_a - I_b * D_a / D_b

This method doesn't consider shadow volumes, which means I asume
the path from camera to the target terrain is always have sunlight
scattered. So there is room for further investigation :/

Other changes along this patch:

the color value is not clamped before saving to scatter texture,
because values that are out of range should be as is for color
substraction, and the clamp operation is done in vertex shader.

Signed-off-by: plysan <plysab@gmail.com>